services/kafka-to-mqtt-bridge/requirements.txt
aiokafka==0.11.0
paho-mqtt==1.6.1
orjson==3.10.7
prometheus-client==0.20.0
